refactor(main): move workflow config to lib
