+++ apps/main/src/app/[locale]/generate/ch/[id]/generation-phases.ts
@@ -16,7 +16,7 @@ import {
   type ChapterStepName,
   type ChapterWorkflowStepName,
   type LessonStepName,
-} from "@/workflows/config";
+} from "@/lib/workflow/config";
 import {
   BookOpenIcon,
   CheckCircleIcon,
