+++ apps/main/src/app/[locale]/generate/cs/[id]/use-generation-phases.ts
@@ -1,12 +1,12 @@
 "use client";
 
 import { type PhaseStatus } from "@/lib/generation-phases";
+import { type CourseWorkflowStepName } from "@/lib/workflow/config";
 import {
   type ThinkingMessageGenerator,
   createCountingGenerator,
   cycleMessage,
 } from "@/lib/workflow/use-thinking-messages";
-import { type CourseWorkflowStepName } from "@/workflows/config";
 import { useExtracted } from "next-intl";
 import {
   PHASE_ICONS,
